separate freebsd specific test code.
